IR Sensor Test 2/19/2015
Use this code to test the IR Sensor.

Hook up the sensor's red wire to the +5V pin on the Arduino board, the
black wire to the GND pin on the board, and the white (signal) wire to
DIGITAL pin 8 on the board.

Ask your teacher for assistance during this process, to make sure
nothing is shorted out.

The sensor should print when it sees different reflective surfaces
(black or white). Test by moving it along the flag.